#9 feat: Result 서비스 고도화
게임서버로부터 받은 메시지 Consume

+++ common/src/main/java/com/ns/common/ClientRequest.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class ClientRequest {
     private Long membershipId;
     private int socket;
-    private int champindex;
+    private Long champindex;
     private String user_name;
     private String team;
     private int channel;
